backend/pyproject.toml
[tool.poetry]
name = "jubileu-backend"
version = "0.1.0"
description = "Backend do clone do Twitter"
authors = ["Nelson Carvalho <seu-email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
